test(ai-openrouter): consolidate finish reason coverage
