+++ packages/ai/openrouter/test/Generated.test.ts
@@ -1,7 +1,9 @@
-import { Generated } from "@effect/ai-openrouter"
+import { Generated, OpenRouterClient, OpenRouterLanguageModel } from "@effect/ai-openrouter"
 import { describe, it } from "@effect/vitest"
 import { deepStrictEqual } from "@effect/vitest/utils"
-import { Schema } from "effect"
+import { Effect, Schema, Stream } from "effect"
+import { type AiError, LanguageModel, type Response, Tool, Toolkit } from "effect/unstable/ai"
+import { HttpClient, HttpClientRequest, HttpClientResponse } from "effect/unstable/http"
 
 describe("Generated", () => {
   it("decodes nullable generation statistics", () => {
@@ -72,4 +74,85 @@ describe("Generated", () => {
 
     deepStrictEqual(Schema.decodeUnknownSync(Generated.ChatUsage)(usage), usage)
   })
+
+  for (
+    const { encrypted, mode, reason, tool } of [
+      { mode: "stream", encrypted: true, tool: true, reason: "tool-calls" },
+      { mode: "stream", encrypted: false, tool: true, reason: "stop" },
+      { mode: "stream", encrypted: true, tool: false, reason: "stop" },
+      { mode: "generate", encrypted: true, tool: true, reason: "tool-calls" }
+    ] as const
+  ) {
+    it.effect(`${mode}: encrypted=${encrypted}, tool=${tool} finishes with ${reason}`, () =>
+      Effect.gen(function*() {
+        const metadata = { id: "response-1", model: "test/reasoning-model", created: 1 }
+        const reasoningDetails = encrypted
+          ? [{ type: "reasoning.encrypted", data: "opaque-signature", format: "unknown" } as const]
+          : []
+        const toolCalls = tool
+          ? [{
+            index: 0,
+            id: "call-1",
+            type: "function" as const,
+            function: { name: "ProbeTool", arguments: "{\"value\":1}" }
+          }]
+          : []
+        const usage = { prompt_tokens: 1, completion_tokens: 1, total_tokens: 2 }
+        const chunks = yield* Schema.decodeUnknownEffect(Schema.Array(Generated.ChatStreamChunk))([
+          {
+            ...metadata,
+            object: "chat.completion.chunk",
+            choices: [{ index: 0, delta: { reasoning_details: reasoningDetails, tool_calls: toolCalls } }]
+          },
+          {
+            ...metadata,
+            object: "chat.completion.chunk",
+            choices: [{ index: 0, delta: {}, finish_reason: "stop" }],
+            usage
+          }
+        ])
+        const body = yield* Schema.decodeUnknownEffect(Generated.SendChatCompletionRequest200)({
+          ...metadata,
+          object: "chat.completion",
+          system_fingerprint: null,
+          choices: [{
+            index: 0,
+            finish_reason: "stop",
+            message: { role: "assistant", reasoning_details: reasoningDetails, tool_calls: toolCalls }
+          }],
+          usage
+        })
+        const response = HttpClientResponse.fromWeb(
+          HttpClientRequest.post("https://example.com/chat/completions"),
+          new globalThis.Response()
+        )
+        const toolkit = Toolkit.make(Tool.make("ProbeTool", {
+          parameters: Schema.Struct({ value: Schema.Number }),
+          success: Schema.String
+        }))
+        const client = OpenRouterClient.OpenRouterClient.of({
+          client: Generated.make(HttpClient.make(() => Effect.die("Unexpected HTTP request"))),
+          createChatCompletion: () => Effect.succeed([body, response]),
+          createChatCompletionStream: () => Effect.succeed([response, Stream.fromIterable(chunks)])
+        })
+        const options = { prompt: "Use the probe tool", toolkit, disableToolCallResolution: true } as const
+        const operation: Effect.Effect<Array<Response.AnyPart>, AiError.AiError, LanguageModel.LanguageModel> =
+          mode === "stream"
+            ? LanguageModel.streamText(options).pipe(Stream.runCollect)
+            : LanguageModel.generateText(options).pipe(Effect.map((result) => result.content))
+        const parts = globalThis.Array.from(
+          yield* operation.pipe(
+            Effect.provide(OpenRouterLanguageModel.model(metadata.model)),
+            Effect.provide(toolkit.toLayer({ ProbeTool: () => Effect.succeed("ok") })),
+            Effect.provideService(OpenRouterClient.OpenRouterClient, client)
+          )
+        )
+
+        deepStrictEqual(
+          parts.filter((part) => part.type === "tool-call").map((part) => part.name),
+          tool ? ["ProbeTool"] : []
+        )
+        deepStrictEqual(parts.filter((part) => part.type === "finish").map((part) => part.reason), [reason])
+      }))
+  }
 })
